LTMU-H/requirements.txt
certifi==2021.5.30
cffi==1.15.0
cycler==0.11.0
Cython==0.29.30
dataclasses==0.8
easydict==1.9
-e git+https://github.com/ddshan/hand_object_detector@27d9577c8e60b9cae3b8eefc0d23262af5e39bfd#egg=faster_rcnn&subdirectory=lib
fire==0.4.0
future==0.18.2
got10k==0.1.3
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
jpeg4py==0.1.4
kiwisolver==1.3.1
lmdb==1.3.0
MarkupSafe==2.0.1
matplotlib==3.3.4
msgpack==1.0.4
numpy==1.19.5
opencv-python==4.6.0.66
pascal-voc-writer==0.1.4
Pillow==8.4.0
protobuf==3.19.4
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==5.3.1
scikit-learn==0.24.2
scipy==1.2.1
Shapely==1.8.4
six==1.16.0
sklearn==0.0
tensorboardX==2.5.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.3.2
torch==1.4.0
torchvision==0.5.0
tqdm==4.19.9
typing_extensions==4.1.1
wget==3.2
yacs==0.1.8
zipp==3.6.0
